backend/requirements.txt
Django==2.2.10
django-ckeditor==5.9.0
django-colorfield==0.1.17
django-js-asset==1.2.2
djangorestframework==3.8.0
pytz==2019.3
sqlparse==0.3.0
wheel==0.24.0
